refactor(shared): breadcrumb component
