+++ components/shared/Breadcrumb/Breadcrumb.tsx
@@ -1,7 +1,30 @@
 import { Children } from "react";
+import Link from "next/link";
 import { cn } from "@/lib/utils";
-import BreadcrumbItem, { BreadcrumbItemProps } from "./BreadcrumbItem";
 import Icon from "../Icon";
+
+export interface BreadcrumbItemProps {
+  /** The displayed text of the breadcrumb item */
+  text: string;
+  /** The href of the breadcrumb item */
+  href?: string;
+  /** A CSS class name for styling the breadcrumb item */
+  className?: string;
+}
+
+export const BreadcrumbItem = ({
+  text,
+  href,
+  className,
+}: BreadcrumbItemProps) => {
+  return href ? (
+    <Link href={href} className={className}>
+      {text}
+    </Link>
+  ) : (
+    <span className={className}>{text}</span>
+  );
+};
 
 export interface BreadcrumbProps {
   /** `Breadcrumb.Item` with text and href */
