+++ components/shared/Breadcrumb/index.tsx
@@ -1,5 +1,2 @@
-import Breadcrumb from "./Breadcrumb";
-
+export { default as Breadcrumb } from "./Breadcrumb";
 export * from "./Breadcrumb";
-
-export default Breadcrumb;
